chore: reorganize helpers used by place order reducer
